Working 1.16 commit!

+++ src/main/java/muramasa/antimatter/recipe/loader/AntimatterRecipeLoader.java
@@ -5,20 +5,51 @@ import muramasa.antimatter.Antimatter;
 import muramasa.antimatter.AntimatterAPI;
 import muramasa.antimatter.recipe.Recipe;
 import muramasa.antimatter.recipe.RecipeMap;
+import net.minecraft.client.Minecraft;
+import net.minecraft.server.MinecraftServer;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.event.TagsUpdatedEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.Collection;
 import java.util.List;
 
-public class AntimatterRecipeLoader {
+public class AntimatterRecipeLoader implements IRecipeRegistrate {
+
     protected List<IRecipeLoader> loaders = new ObjectArrayList<>();
 
+    boolean vanillaLoaded = false;
+    boolean customLoaded = false;
+    boolean loadedStart = false;
+
+    @Override
+    public void add(IRecipeLoader load) {
+        loaders.add(load);
+    }
+
     @SubscribeEvent
-    public void TagsUpdatedEvent(TagsUpdatedEvent event)
+    public void TagsUpdatedEvent(final TagsUpdatedEvent.VanillaTagTypes event)
     {
+        vanillaLoaded = true;
+        load();
+    }
+
+    @SubscribeEvent
+    public void TagsUpdatedEvent(final TagsUpdatedEvent.CustomTagTypes event)
+    {
+        customLoaded = true;
+        load();
+    }
+
+    private void load() {
+        if (!(vanillaLoaded && customLoaded)) return;
+        AntimatterAPI.all(RecipeMap.class, RecipeMap::compile);
+    }
+
+    public void loadRecipes() {
+        if (loadedStart) return;
+        loadedStart = true;
         long time = System.currentTimeMillis();
-        AntimatterAPI.all(RecipeMap.class, RecipeMap::reset);
         loaders.forEach(IRecipeLoader::init);
         Antimatter.LOGGER.info("Time to load all AM recipes: " + (System.currentTimeMillis()-time) + " ms");
         long recipes = AntimatterAPI.all(RecipeMap.class).stream().mapToLong(rm -> {
@@ -26,9 +57,5 @@ public class AntimatterRecipeLoader {
             return rs == null ? 0 : rs.size();
         }).sum();
         Antimatter.LOGGER.info("Total recipes " + recipes);
-    }
-
-    public void registerRecipeLoader(IRecipeLoader loader) {
-        loaders.add(loader);
     }
 }
